행렬곱에 사용되는 변수 타입 int에서 float로 수정

+++ Survivor/Survivor/Math.cpp
@@ -47,7 +47,7 @@ Matrix3 operator*(const Matrix3& lhs, const Matrix3& rhs)
 	// 캐시 효율을 위해 rkc 순으로 연산
 	for (int r = 0; r < 3; r++) {
 		for (int k = 0; k < 3; k++) {
-			int temp = lhs.mat[r][k];
+			float temp = lhs.mat[r][k];
 
 			for (int c = 0; c < 3; c++) {
 				ret.mat[r][c] += temp * rhs.mat[k][c];
@@ -113,7 +113,7 @@ Matrix4 operator*(const Matrix4& lhs, const Matrix4& rhs)
 	// 캐시 효율을 위해 rkc 순으로 연산
 	for (int r = 0; r < 4; r++) {
 		for (int k = 0; k < 4; k++) {
-			int temp = lhs.mat[r][k];
+			float temp = lhs.mat[r][k];
 
 			for (int c = 0; c < 4; c++) {
 				ret.mat[r][c] += temp * rhs.mat[k][c];
